Split Transpiler tests
